+++ src/main/java/com/example/demo/dto/ProdutoResponseDto.java
@@ -1,6 +1,5 @@
 package com.example.demo.dto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 import com.example.demo.enums.TipoProduto;
@@ -13,5 +12,5 @@ public class ProdutoResponseDto {
 	private UUID id;
 	private String nome;
 	private TipoProduto tipo;
-	private BigDecimal precoUnitario;
+	private Double precoUnitario;
 }
